test/system: deflake podman healtchcheck test
Using sleeps is by definition not robust. Unfortunately we want to
actually test the healthcheck timer fires in the background so we have
to wait as such there is no way around waiting. However one trick we can
use instead of sleeping and doing nothing we can (ab)use podman events
until --until which will list and wait events until the given time.
With that we can read and then directly check events without having to
do both sleep and events separately.

Hopefully this is enough to make the test stable.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>