Release docs: add a way to bypass the CI skips
Background: As of #23174, CI is selectively run depending on
the set of files changed. Changes to version.go should require
a full run of all tests, because we want releases tested.

Problem: RC releases bump version.go and then bump it back.
Cirrus's changesInclude() sees this as a NOP, so it does
not trigger the "run all CI tests" rule.

Solution: require CI:ALL in release PRs. Add step to process.

Signed-off-by: Ed Santiago <santiago@redhat.com>