[v4.4.1-rhel] Change images_test expect from 10 to 9
I'm not exactly sure what is happening here, but this call
```
result := podmanTest.Podman([]string{"images", "-q", "-f", "reference=quay.io/libpod/*"})
```
in the test/e2e/images-test.go in this test
```
It("podman images filter reference", func()
```

is now sending back 10 instead of 9 objects.  This was a change
that @edsantiago also made in https://github.com/containers/podman/pull/21356

After the other adjustments I made to the tests to right them,
this seemed to be the last issue.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>