[v5.2] update buildah-bud buildah-tests.diff
When trying to vendor Buildah v1.37.0, all kinds of mayhem
ensued.  I followed https://github.com/containers/podman/tree/main/test/buildah-bud
and the result was this diff file.

I'm not sure if this is right, and as I'm in a very deep dark corner
of the project that I've not been in before, I'll ask @edsantiago
to throw the life ring if necessary.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>