[v4.9-rhel] libpod: bind ports before network setup
We bind ports to ensure there are no conflicts and we leak them into
conmon to keep them open. However we bound the ports after the network
was set up so it was possible for a second network setup to overwrite
the firewall configs of a previous container as it failed only later
when binding the port. As such we must ensure we bind before the network
is set up.

This is not so simple because we still have to take care of
PostConfigureNetNS bool in which case the network set up happens after
we launch conmon. Thus we end up with two different conditions.

Also it is possible that we "leak" the ports that are set on the
container until the garbage collector will close them. This is not
perfect but the alternative is adding special error handling on each
function exit after prepare until we start conmon which is a lot of work
to do correctly.

For upstream:
Fixes https://issues.redhat.com/browse/RHEL-50746 for upstream

For RHEL 4.9-rhel/RHEL 8.10/9.4:
Fixes https://issues.redhat.com/browse/ACCELFIX-302, https://issues.redhat.com/browse/RHEL-62522

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit 77081df8cdf0d2a68d52eebacc35d3e6a1df7660)
Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>