CI: podman-machine: do not use cache registry
It's too difficult to keep the podman-machine image up-to-date.
And, we can't use the cache on Mac/Windows, so if quay is down
we're hosed no matter what.

Add a "nocache" mechanism to install_test_configs() and use that
in machine test setup.

Manual cherrypick of #23461 from main, one trivial conflict.

Signed-off-by: Ed Santiago <santiago@redhat.com>