podman.spec: `Requires: composefs`
This has come up multiple times, most recently in
https://issues.redhat.com/browse/RHELMISC-4891

But also previously when we tried to enable composefs by default
in rawhide for a while it was belatedly realized that we were
missing this implicit runtime dependency.

Since composefs is tiny, and we want to promote it and make
it easier to use, and we consider it a core target for RHEL10,
let's add a hard requirement by default.

Signed-off-by: Colin Walters <walters@verbum.org>