Specify format to use for referencing fixed bugs.
Per discussion within the team, we would like all bugfixes to be
referenced in commit messages as "Fixes:" (or "Fixes" with no
colon, for those who would like to type one less character). We
are going to add CI to enforce this on our LTS branches as we
should only be backporting bugfixes there.

Signed-off-by: Matt Heon <mheon@redhat.com>