cirrus: remove 3rd party connectivity check
This doesn't help us at all, first the list is outdated. AFAICT we no
longer connect to docker.io, registry.fedoraproject.org or
podman.cachix.org (seems to be a cache site for nix.dev?) anywhere in
our tests.

Second a simple port check is not helpful, in the most cases the
CDN's and or load balancer accept connections but return internal server
errors when the registy goes down.

This is very similar to commit 5b6de98ee8.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>