cirrus: move renovate check into validate
The renovate config is used for the renovate bot, validating this in the
prior fedora prebuild setp is just confusing and hidden.

The problem is this image is very big so it is slow to download/extract.
To speed things up given it is only a single file we check the diff if
we even changed it.

Now one could argue this should be part of the validate Makefile target
but I given the size I do not want this run by default and I am not sure
if we should do the diff check in the Makefile.

Lastly remove -it, these is meant for interactive use and throws a
warning here because we have no actual tty attached.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>