github: remove fcos-podman-next-build-prepush
It is flaking[1] from time to time on PRs and doesn't really add value
because noone is actually chaing the Containerfile. The task on main
that build a image based of main still exists as per Lokesh
podman-dekstop uses it for testing. In the near term we will get proper
builds in the podman-machine-os repo so this other workflow can
hopefully removed altogether.

[1] https://github.com/containers/podman/actions/runs/10618524888/job/29433963845?pr=23807

Signed-off-by: Paul Holzinger <pholzing@redhat.com>