cirrus: test only on f40/rawhide
We want to update to golang 1.22 but f39 only ships 1.21. This means we
can no longer build on f39 once the update is done. Given we do not ship
podman 5.0 on f39 anyway this is not a problem.

So drop the f39 tasks but given we use them to test different db/storage
drivers we cannot just remove them. Instead we now test f40 twice with
the different db/storage drivers to ensure we keep full coverage.

It do however drop it from the container int task. I really do not see
the value there to test a different db backend as this doesn't depend on
anything in the container.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>