Packit: reorg Fedora and ELN targets
This commit separates out ELN copr jobs into a separate set so that
other Fedora targets can be reused with yaml anchors. Also, given ELN is not
really Fedora, but rather a RHEL-next, it only makes sense to separate it out this way.

The Fedora copr and downstream targets also mention fedora-40 explicitly
because once Fedora-41 is released, `fedora-latest` and
`fedora-latest-stable` will both point to `fedora-41` , so we need
to include fedora-40 as well.

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>